tests: add "ApMetricsResponseR3" test in boardfarm
Added R3 version of "ApMetricsResponse" test in boardfarm.
Modified existing "send_and_check_policy_config_metric_reporting"
function to support "Associated Wifi 6 Sta Status Report Inclusion
Policy" bit in "Metric Reporting Tlv".

PPM-1696.

+++ common/beerocks/bwl/include/bwl/mon_wlan_hal_types.h
@@ -15,7 +15,7 @@
 namespace bwl {
 
 const int WIFI_SSID_MAX_LENGTH = 32 + 1 + 3; // 1 byte for null termination + 3 for alignment
-constexpr int TID_UP_MAX       = 8;          // Maximum number of user priorities
+constexpr uint8_t TID_UP_MAX   = 8;          // Maximum number of user priorities
 
 struct SMacAddr {
     uint8_t oct[6];
